Use mock validator more aggressively
